Add ability to use IAM role
Omit the access key and secret key from s3_dir and it will fall back on
the IAM Profile attached to the EC2 instance.

This is much more secure and follows best practices in aws.

This required default parameters on the S3Lib initialize call.

In order to avoid changing the order of arguments on the initializer,
the region and is_mock properties now have defaults.

Closes #6