Add ability to read IAM role from server if running in aws. Also remove change of file mode on files.
1. Ideally we should be able to omit the acces key and secret key from the call to s3_dir and it should fall back on the role
that the host uses. This is much more secure and follows best practices in aws. This is fairly easy to do but required some default parameters
on the S3Lib initialiaze call. Due to those being first the the method signature I had to change the region and is_mock to defaults as well. I thought
this is preferable to changing the order. Open to suggestions.

2. THere was some strange functionality around adding the execute bit to the file mode. Im guessing this is to allow recursive copying in the case of nested directories. However it had two flaws
    a. It was adding executable bits to every file not just to directories.
    b. there was a call to mode.to_i.even? that was taking into account the entire mode rather than just the number it was talking about. In the case of an 0700 mode it would mean that it would be bumped up to
       0800 and be an invalid file mode.
given both of these problems I took the loop out, But im also happy to find ways to add this back in.